+++ app/assets/javascripts/code_listing/annotation.ts
@@ -1,71 +1,72 @@
 import { CodeListing } from "code_listing/code_listing";
-import { SuperAnnotation, AnnotationType } from "code_listing/super_annotation";
+export type AnnotationType = "user" | "error" | "warning" | "info";
 
-const ORDERING = ["error", "warning", "info"];
 
-export interface AnnotationData {
-    type: AnnotationType;
-    text: string;
+export abstract class Annotation {
+    readonly id: number;
+    protected shown: boolean = true;
+
+    protected annotation: HTMLDivElement;
+    protected dot: HTMLSpanElement;
+
+    protected readonly codeListingHTML: HTMLTableElement;
+    protected readonly codeListing: CodeListing;
+
     row: number;
-}
+    type: AnnotationType;
 
-export class Annotation extends SuperAnnotation {
-    readonly text: string;
-
-    constructor(id: number, m: AnnotationData, listing: HTMLTableElement, codeListing: CodeListing) {
-        super(id, listing, codeListing, m.row + 1, m.type);
-        this.type = m.type;
-        this.text = m.text;
-        this.createHTML();
+    protected constructor(id: number, codeListingHTML: HTMLTableElement, codeListing: CodeListing, row: number, type: AnnotationType) {
+        this.id = id;
+        this.codeListingHTML = codeListingHTML;
+        this.codeListing = codeListing;
+        this.row = row;
+        this.type = type;
     }
 
-    protected createAnnotation(): void {
-        let annotationRow: HTMLTableRowElement = this.codeListingHTML.querySelector(`#annotations-${this.row}`);
-        if (annotationRow === null) {
-            annotationRow = this.createAnnotationRow();
+    protected createHTML(): void {
+        this.createAnnotation();
+        this.createDot();
+    }
+
+    hide(): void {
+        this.annotation.classList.add("hide");
+        this.shown = false;
+        this.addDot();
+    }
+
+    show(): void {
+        this.annotation.classList.remove("hide");
+        this.shown = true;
+        this.removeDot();
+    }
+
+    protected abstract createAnnotation(): void;
+
+    protected createDot(): void {
+        const codeGutter = this.codeListingHTML.querySelector(`tr#line-${this.row} .rouge-gutter.gl`);
+        const potentialDot = codeGutter.querySelector("span.dot") as HTMLSpanElement;
+        if (potentialDot !== null) {
+            this.dot = potentialDot;
+            return;
         }
 
-        this.annotation = document.createElement("div");
-        this.annotation.setAttribute("title", this.type[0].toUpperCase() + this.type.substring(1));
-        this.annotation.classList.add("annotation", this.type, "machine-annotation");
-        this.annotation.appendChild(document.createTextNode(
-            this.text.split("\n").filter(s => !s.match("^--*$")).join("\n")
-        ));
+        this.dot = document.createElement("span");
+        this.dot.setAttribute("class", `dot dot-${this.type}`);
 
-        const edgeCopyBlocker = document.createElement("div");
-        edgeCopyBlocker.setAttribute("class", "copy-blocker");
+        const titleAttr = I18n.t("js.annotation.hidden");
+        this.dot.setAttribute("title", titleAttr);
 
-        let annotationGroup: HTMLDivElement = annotationRow.querySelector(`.annotation-cell .annotation-group-${this.type}`);
-        if (annotationGroup == null) {
-            this.createAnnotationGroups(annotationRow.querySelector(".annotation-cell"));
-            annotationGroup = annotationRow.querySelector(`.annotation-cell .annotation-group-${this.type}`);
-        }
-        annotationGroup.appendChild(this.annotation);
-        annotationGroup.appendChild(edgeCopyBlocker);
+        codeGutter.prepend(this.dot);
     }
 
-    private createAnnotationRow(): HTMLTableRowElement {
-        const codeRow: HTMLTableRowElement = this.codeListingHTML.querySelector(`#line-${this.row}`);
-        const annotationRow: HTMLTableRowElement = this.codeListingHTML.insertRow(codeRow.rowIndex + 1);
-        annotationRow.setAttribute("class", "annotation-set");
-        annotationRow.setAttribute("id", `annotations-${this.row}`);
-        const htmlTableDataCellElement = annotationRow.insertCell();
-        htmlTableDataCellElement.setAttribute("class", "rouge-gutter gl");
-        htmlTableDataCellElement.appendChild(document.createElement("div"));
-        const annotationTDC: HTMLTableDataCellElement = annotationRow.insertCell();
-        annotationTDC.setAttribute("class", "annotation-cell");
-
-        this.createAnnotationGroups(annotationTDC);
-
-        return annotationRow;
+    protected addDot(): void {
+        this.dot.classList.add(`dot-${this.type}`);
     }
 
-    private createAnnotationGroups(annotationTDC: HTMLTableDataCellElement): void {
-        for (const type of ORDERING) {
-            const groupDiv: HTMLDivElement = document.createElement("div");
-            groupDiv.setAttribute("class", `annotation-group-${type}`);
-            annotationTDC.appendChild(groupDiv);
+    protected removeDot(): void {
+        const allHiddenOfThisType = this.codeListing.getAnnotationsForLine(this.row).filter(m => m.type === this.type).every(m => m.shown);
+        if (allHiddenOfThisType) {
+            this.dot.classList.remove(`dot-${this.type}`);
         }
     }
 }
-
